+++ packages/e2e/tests/apps/product-feed/assertions/assert-app-render.ts
