+++ packages/e2e/tests/apps/product-feed/operations/navigate-to-category-mapping.ts
