+++ packages/e2e/tests/apps/product-feed/operations/set-category-mapping.ts
