+++ packages/e2e/tests/assertions/assert-app-available.ts
@@ -1,6 +1,6 @@
 import { expect, Page } from "@playwright/test";
-import { routing } from "../setup/routing";
-import { configuration } from "../setup/configuration";
+import { routing } from "../../setup/routing";
+import { configuration } from "../../setup/configuration";
 
 interface checkIfAppIsAvailableArgs {
   page: Page;
