+++ packages/e2e/tests/operations/install-app.ts
@@ -1,5 +1,5 @@
 import { Page } from "@playwright/test";
-import { routing } from "../setup/routing";
+import { routing } from "../../setup/routing";
 
 interface InstallTheAppArgs {
   page: Page;
